+++ lib_paysages/auto.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 void autoSetDaytime(int hour, int minute);
-void autoSetDaytimeFraction(double daytime);
+void autoSetDaytimeFraction(float daytime);
 void autoGenRealisticLandscape(int seed);
 
 #ifdef __cplusplus
